introduce a `TxGenerationResult`
We now keep track of tx generation events, instead of an `Either` converted to and `Option` in `withTxGenerationLog`. This new class replaces `Option`s in `*CommitPublished` structures (*) and makes its way all the way to the serialization.

NB: There has been a changed in the compat logic in the codecs compared to previous commits.

(*) only in `LocalCommitPublished` for now.